features: strict upgrade checks on startup
This check should prevent a user inadvertently damaging
their cluster by trying to skip many feature releases
during an upgrade, which might skip important data migration
or compatibility code.

This is not an ultra-strict check that the user is only
hopping one logical version: we will tolerate upgrades
from as far back as the `earliest_logical_version`, which
is currently set to 7 (Redpanda v22.3.1)

In special circumstances, this check may be skipped by setting
the `upgrade_override_checks` node configuration property to true.