rpc: update rpc_gen_cycling_test
Tests that relate to version upgrade are removed, as
we no longer support a range of versions.

Tests that used ADL now use serde.  The mock request/response
classes that specified a particular encoding are gone in favor
of using serde everywhere.

The encode_for_version method now needs to pass through
versions for v2 transports because we have to use a v2
transport for testing what happens when a client uses
a too-new version.