serde/serde.h: added support for std::array as size,elements (+ throw)
serialization is done as std::vector,
deserialization will check that size matches the std::tuple_size<std_array_t>,
and will throw if different.
This is done as a safety measure in case of a size change without a change in decoding logic of old binary data.
It's meant to catch bugs during testing, not as a recoverable error.