archival: do fast manifest upload if interval is disabled
If there is no interval set, then interpret this as
"always upload immediately": do the upload after
segment uploads (this is the <=23.1 behavior) and
skip the upload we otherwise might do in parallel
with segment uploads.

This is useful for fixture tests that like to make
very specific assertions about upload counts, and
would be awkward if they had to account for
walltime-based interval uploads.