tests: fix verifiable_consumer get_committed-offsets
In commit 63a17d5c this class was modified in an inconsistent
way, the last_committed timestamp is now read as if it were
a callable: this causes an exception when tests fail (this is
the only time anything will call get_committed_offsets).

Fix this by removing the unused last_committed attribute, and
having get_committed_offsets return a dict of worker to the
worker's committed offset map.