cluster: add cloud_storage_usage controller RPC
A new controller RPC is introduced in this commit:
`cloud_storage_usage`. The request takes a list of partitions broken
down by shard, and the response contains the total number of bytes used
by the cloud storage log of the specified partitions and the partitions
that could not be found. This RPC is a building block for cluster wide
map-reduce operation that computes the total size of all cloud logs.