admin: Add blocked_reactor_notify_ms command
The command changes the timeout of stall detector for brief period of
time.

curl -XPUT  "127.0.0.1:9644/v1/config/blocked_reactor_notify_ms/1?expires=15"

This command will change the value to 1ms and it will be reset back
after 15 seconds. If 'expires' parameter is not set the default expire
value will be used. The default is 5 minutes.