test_utils: sleep in debug mode in flush_tasks
Previously, we slept but sometimes saw that the sleep
was insufficient: this was addressed in
https://github.com/redpanda-data/redpanda/pull/8464

However, it appears that sometimes in debug mode tests
can still not see the ordering they expect, probably
because of the scheduler's shuffle().  So in debug
mode, reinstate the simple sleep(10ms), in addition
to the inter-CPU task calls that aim to flush preceding
work.