c/members_manager: include all nodes still in raft0 in kvstore snapshot
If a restart happens when a node was removed (i.e. the removed command was
processed) but is still in the controller group, we may still need to
recreate a connection to that node after restart. This means that we
need to track a set of these nodes and include them in the kvstore
snapshot.