archival: fix re-upload of nop compacted segments
This commit fixes a bug in the compacted segment reupload logic.

Previously, if the compaction of a local segment resulted in the segment
maintaing its size (i.e. all the keys in the segment were unique) the
following sequence of events would occur:
1. The re-upload logic finds such a segment selects such a segment as
   the sole re-upload candidate.
2. The segment is re-uploaded, with the same name as previously. The
   size of the segment is part of the addressable name in cloud storage,
   in order to distinguish between original and re-uploaded segments.
3. The original segment is deleted. Since both the original and
   re-uploaded segments are addressed by the same name, the segment is
   entirely lost.

This can lead to gaps in the cloud log, which, in turn, prevent the read
path from making progress (consumers stall).

The fix is to update the logic which picks the range of segments to be
re-uploaded ('archival::segment_collector') to ignore re-uploads that
would lead to name clashing. Note that this can only be the case when
only one segment is picked for re-upload. If more segments are selected,
the size is guaranteed to change, a name clash is not possible.