k/consumer: pass result of name() to on_stopped
Previously, the value of `_name` was passed into `on_stopped`. This is a
problem in the case of an unnamed consumer because `_name` will be an
empty string. Instead, use the result of `consumer::name()`.

Here is the definition of `on_stopped`:
```cpp
auto on_stopped = [this, group_id](const member_id& name) {
    _consumers[group_id].erase(name); // silenty fails if name is empty
};
```
The result of passing an empty string to `on_stopped` is that the
consumer is not erased from the map. This is an issue because the client
expects the map to eventually become empty.

From `client::stop()`:
```cpp
for (auto& [id, group] : _consumers) {
    while (!group.empty()) { // group is never empty, so infinite loop
        auto c = *group.begin();
        co_await catch_and_log([c]() {
            return c->leave();
        });
    }
}
```

Instead, the while loop becaomes an inifite loop since the consumer is
never removed.

Fixes #9310