rptest: relax cloud storage usage checks
This commit tweaks the CloudStorageUsage tests to maintain a sliding
window of the reported usages and check if the usage inferred from the
manifest is present in the window. Even after the previous fixes, we've
seen failures due to the uploaded manifest lagging behind.

The size of the sliding window allows the test to "look back" in time by
2 seconds. Manifest uploads are performed every 1 second, so this should
be more than enough.