Increase OMB start timeout to 5 minutes
This is the timeout from when the benchmark has been started to
when the warmup traffic starts. Since this also involves deleting the
topics from the previous run (because reset: true), this may take a
long time as topic creation and user group formation can be greatly
slowed down by the work associated with the deletion.

This causes tests which run after max_partitions_test to often fail to
start. Later we may get a more sophisticated cleanup strategy, but for
now just increase the start timeout to 5 minutes.