Check that we have non-zero warmup time
To detect test start we look for a "starting warm-up traffic" message
which only appears if the test has > 0 warmup time, otherwise we will
just hang for 5 minutes and then time out the startup (even though
the test is likely still running).

Instead, assert that warmup time is at least 1.