crash_tracker: define API
Define the core API of the crash recording module ahead of development
to help parallelise work by reducing dependencies and unblocking some
tasks.

Note though the dependencies between the components, the types used in
the interfaces and the lifetimes of the objects. Specifically note that
the `recorder` is defined as a static global singleton. This is to
ensure that it can be used in signal handler functions which are defined
as function pointers.