r/append_entries_buffer: added metrics to the buffer
Added metrics allowing estimating the size of the append entries buffer.
This may be helpful to validate how well Raft coalesce writes and
flushes. The average size of the buffer can be calculated as the ratio
between the number of append entries requests and append entries buffer
flushes executed in the period of time.

Signed-off-by: Michał Maślanka <michal@redpanda.com>